Effects can be turned on and off while lab is running.
Implemented an action for each of the power buttons; if the lab is running, clicking a power button will cause the synth chain to be rebuilt.

Also, effectObjects and effectSynths are initialised to arrays of nil in the init method, and are reset to arrays of nil in the freeAll method.