Made changes to allow EffectUnits to persist when their synths are freed.
addEffect/removeEffect are used respectively to add and remove EffectUnits to and from the chain and will not start synths.

startEffect/stopEffect are used to control which synths are running.

Other minor changes:
storeIOSynthDefs is now a class method.
createChain is now createEmptyChain.
freeAll now frees the synth group.
init method stripped of redundant code.