Added the findGUI class method.
Instead of calling EffectUniGUI.new(effectUnit), EffectUnitGUI.findGUI(effectUnit) can be called - this will attempt to find an existing GUI for the EffectUnit. If a GUI is found, it will be moved to the front, otherwise a new GUI will be created.

This means that duplicate GUIs for the same EffectUnit do not need to be created (but they can be created, by calling EffectUnitGUI.new directly).